# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string()
})


/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "slug": zod.string(),
  "price": zod.number(),
  "originalPrice": zod.number().nullish(),
  "category": zod.string(),
  "description": zod.string(),
  "image": zod.string().nullish(),
  "badge": zod.string().nullish(),
  "rating": zod.number(),
  "reviewCount": zod.number(),
  "inStock": zod.boolean(),
  "isNew": zod.boolean(),
  "isBestseller": zod.boolean(),
  "colors": zod.array(zod.string()),
  "features": zod.array(zod.string()),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})
export const ListProductsResponse = zod.array(ListProductsResponseItem)


/**
 * @summary Create a product
 */
export const CreateProductBody = zod.object({
  "name": zod.string(),
  "slug": zod.string(),
  "price": zod.number(),
  "originalPrice": zod.number().nullish(),
  "category": zod.string(),
  "description": zod.string(),
  "image": zod.string().nullish(),
  "badge": zod.string().nullish(),
  "rating": zod.number(),
  "reviewCount": zod.number(),
  "inStock": zod.boolean(),
  "isNew": zod.boolean(),
  "isBestseller": zod.boolean(),
  "colors": zod.array(zod.string()),
  "features": zod.array(zod.string())
})


/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  "id": zod.coerce.number()
})

export const UpdateProductBody = zod.object({
  "name": zod.string(),
  "slug": zod.string(),
  "price": zod.number(),
  "originalPrice": zod.number().nullish(),
  "category": zod.string(),
  "description": zod.string(),
  "image": zod.string().nullish(),
  "badge": zod.string().nullish(),
  "rating": zod.number(),
  "reviewCount": zod.number(),
  "inStock": zod.boolean(),
  "isNew": zod.boolean(),
  "isBestseller": zod.boolean(),
  "colors": zod.array(zod.string()),
  "features": zod.array(zod.string())
})

export const UpdateProductResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "slug": zod.string(),
  "price": zod.number(),
  "originalPrice": zod.number().nullish(),
  "category": zod.string(),
  "description": zod.string(),
  "image": zod.string().nullish(),
  "badge": zod.string().nullish(),
  "rating": zod.number(),
  "reviewCount": zod.number(),
  "inStock": zod.boolean(),
  "isNew": zod.boolean(),
  "isBestseller": zod.boolean(),
  "colors": zod.array(zod.string()),
  "features": zod.array(zod.string()),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})


/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  "id": zod.coerce.number()
})