# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  slug: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  description: string;
  image?: string | null;
  badge?: string | null;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isNew: boolean;
  isBestseller: boolean;
  colors: string[];
  features: string[];
}